adding all libs to the main file
