+++ .vscode/settings.json
@@ -2,5 +2,5 @@
     "files.associations": {
         "sistematriangularsuperior.h": "c"
     },
-    "C_Cpp.errorSquiggles": "Disabled"
+
 }
